бэкэнд 2.0 исправил ошибки

+++ controllers/movies.js
@@ -1,8 +1,14 @@
+// controllers/movies.js
 const Movie = require('../models/movie');
 const handleError = require('../utils/errorHandler');
 
 const getSavedMovies = (req, res) => {
   const userId = req.user._id;
+  // Валидация: проверка, существует ли у пользователя _id
+  if (!userId) {
+    const validationError = { status: 400, message: 'Некорректный идентификатор пользователя' };
+    return handleError(validationError, res);
+  }
   Movie.find({ owner: userId })
     .then((movies) => res.send(movies))
     .catch((err) => handleError(err, res));
@@ -13,6 +19,12 @@ const createMovie = (req, res) => {
     country, director, duration, year, description, image, trailerLink, nameRU,
     nameEN, thumbnail, movieId,
   } = req.body;
+  // Проверка наличия всех обязательных полей
+  if (!country || !director || !duration || !year || !description
+    || !image || !trailerLink || !nameRU || !nameEN || !thumbnail || !movieId) {
+    const validationError = { status: 400, message: 'Все поля обязательны для заполнения' };
+    return handleError(validationError, res);
+  }
   const userId = req.user._id;
   Movie.create({
     country,
@@ -35,6 +47,11 @@ const createMovie = (req, res) => {
 const deleteMovie = (req, res) => {
   const { _id } = req.params;
   const userId = req.user._id;
+  // Валидация: проверка, существует ли у пользователя _id
+  if (!userId) {
+    const validationError = { status: 400, message: 'Некорректный идентификатор пользователя' };
+    return handleError(validationError, res);
+  }
   Movie.findOneAndDelete({ _id, owner: userId })
     .then((movie) => {
       if (!movie) {
